+++ locust/__init__.py
@@ -1,4 +1,4 @@
-from core import Locust, WebLocust, SubLocust, require_once, task
+from core import Locust, WebLocust, SubLocust, task
 from exception import InterruptLocust, ResponseError, RescheduleTaskImmediately
 
 version = "0.5.1"